requirements.txt
opencv-python
opencv-contrib-python
numpy
pyyaml
# pyrealsense2
pyzmq